+++ src/relay/pty-shell-utils.ts
@@ -10,11 +10,11 @@ import {
 } from '../shared/agent-process-recognition'
 import { getFirstCommandToken } from '../shared/command-token-scanner'
 import {
-  getProcessTableIndex,
   getProcessTableSnapshot,
   type ProcessTableIndex,
   type ProcessTableRow
 } from '../shared/process-table-snapshot'
+import { getProcessTableIndex } from '../shared/process-table-index'
 import { selectForegroundProcessCandidate } from '../shared/foreground-process-selection'
 import {
   resolveOuterWrapperForegroundProcess,
